refactor(initial-board): InitialBoard -> BoardFactory 내부로 이동
